+++ R/app_ui.R
@@ -19,7 +19,7 @@ app_ui <- function(request) {
       title = tags$span(
         "MetMiner",
         # Add the version as a subscript.
-        tags$sub("version 2.0.0",
+        tags$sub("version 2.0.1",
                  style = "font-size: 0.75rem; color: #6c757d; font-weight: normal; margin-left: 5px; letter-spacing: 0;"),
         class = "fw-bold",
         style = "font-size: 1.6rem; color: #008080; letter-spacing: -0.5px; vertical-align: middle;"